requirements-dev.txt
pytest_django==3.1.2
pytest==3.1.2
py==1.4.34
flake8==3.3.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.5.0
model_mommy==1.3.2
pytest_mock==1.6.0

ipython[notebook]

-r requirements.txt